Add C# build on focus
This commit adds a new `automatic_build` setting that, when set to OnFocus, allows GodotTools to automatically build the C# project when any .cs files have been changed by an external IDE after refocusing on the Godot editor. When set to Always, the periodic scan for Assembly Reload will also rebuild the project, allowing compilation even if Godot is not focused when a script change is detected.